chore: unify Makeswift component registrations
