+++ .changeset/eleven-shoes-unite.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Unify Makeswift component registrations
