+++ core/components/footer/index.tsx
@@ -16,7 +16,7 @@ import { client } from '~/client';
 import { readFragment } from '~/client/graphql';
 import { revalidate } from '~/client/revalidate-target';
 import { logoTransformer } from '~/data-transformers/logo-transformer';
-import { SiteFooter as FooterSection } from '~/lib/makeswift/components/site-footer/site-footer';
+import { SiteFooter as FooterSection } from '~/lib/makeswift/components/site-footer';
 
 import { FooterFragment, FooterSectionsFragment } from './fragment';
 import { AmazonIcon } from './payment-icons/amazon';
