+++ core/lib/makeswift/components/customer-group-slot/client.tsx
