+++ core/lib/makeswift/components/product-card/client.tsx
