+++ core/lib/makeswift/components/products-carousel/client.tsx
