+++ core/lib/makeswift/components/products-list/client.tsx
