+++ core/lib/makeswift/components/section/register.ts
