+++ core/lib/makeswift/components/sticky-sidebar/register.ts
